made cahnges to UI and moves styling into styles folder
