+++ src/styles/Hamburger.css
@@ -1,0 +1,66 @@
+.hamburger-menu {
+  /* display: none; */
+  z-index: 9;
+  top: 0;
+  left: 0;
+  right: 0;
+  position: fixed;
+  height: 100%;
+  width: 100%;
+}
+
+.menu-layer {
+  position: relative;
+  /* background-color: #e3f0ff; */
+  background-color: #000000;
+  height: 100%;
+  overflow: hidden;
+}
+
+.menu-links {
+  /*  position and top:
+  add spacing to the top of page and links */
+  position: relative;
+  top: 250px;
+}
+
+nav {
+  display: block;
+  /* text align: moves links to center of page
+  if not here then links positioned at left side of page */
+  text-align: center;
+}
+
+ul {
+  /* not sure what these two do
+  when i remove them the lonks move slightky to the left and */
+  margin: 0;
+  padding: 0;
+}
+
+li {
+  list-style: none;
+  font-size: 6rem;
+  font-weight: 700;
+  height: 135px;
+  position: relative;
+}
+
+/* link text = white, there is no line under links */
+a {
+  color: #fff;
+  text-decoration: none;
+}
+
+/* text turns black when hovering */
+
+.home:hover {
+  color: #fba24f;
+}
+
+.about:hover {
+  color: #5b59f3;
+}
+.experience:hover {
+  color: #ff7e5b;
+}
