+++ src/styles/Home.css
@@ -1,0 +1,49 @@
+@keyframes gradient {
+  0% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+  100% {
+    background-position: 0% 50%;
+  }
+}
+
+.home-container {
+  /* background-image: url("./assets/gradient.jpg");
+  background-repeat: no-repeat;
+  background-size: 100% 100%; */
+  background: linear-gradient(-45deg, #ee7752, #e73c7e, #23a6d5, #23d5ab);
+  background-size: 400% 400%;
+  animation: gradient 15s ease infinite;
+  top: 0;
+  left: 0;
+  right: 0;
+  position: fixed;
+  height: 100%;
+  width: 100%;
+}
+
+.home-wrapper {
+  height: 100%;
+  font-size: 6rem;
+  text-align: center;
+}
+
+p {
+  color: #ffffff;
+  /* font-family: "Shrikhand"; */
+  font-family: "Courier New", monospace;
+  position: relative;
+  top: 200px;
+}
+
+.home-text p {
+  margin-top: 0;
+}
+
+.welcome p {
+  font-size: 1.4rem;
+  font-family: "Gill Sans", sans-serif;
+}
